Move Review app `goToExample()` tests to app package
This change groups Review app “Other examples” templates and tests together so the `govuk-frontend` package no longer relies on the Review app

See GDS Way “Project directory structure”
https://gds-way.cloudapps.digital/manuals/programming-languages/nodejs/#project-directory-structure
