+++ packages/govuk-frontend/src/govuk/components/checkboxes/checkboxes.puppeteer.test.js
@@ -1,5 +1,4 @@
 const {
-  goToExample,
   getAttribute,
   getProperty,
   isVisible,
@@ -290,61 +289,6 @@ describe('Checkboxes with multiple groups and a "None" checkbox and conditional 
   })
 
   describe('when JavaScript is available', () => {
-    /** @type {globalThis.page} */
-    let page
-
-    let $inputsPrimary
-    let $inputsSecondary
-    let $inputsOther
-
-    beforeEach(async () => {
-      page = await goToExample(browser, 'conditional-reveals')
-
-      $inputsPrimary = await page.$$(
-        '.govuk-checkboxes__input[id^="colour-primary"]'
-      )
-      $inputsSecondary = await page.$$(
-        '.govuk-checkboxes__input[id^="colour-secondary"]'
-      )
-      $inputsOther = await page.$$(
-        '.govuk-checkboxes__input[id^="colour-other"]'
-      )
-    })
-
-    it('none checkbox unchecks other checkboxes in other groups', async () => {
-      // Check some checkboxes in the first and second groups
-      await $inputsPrimary[2].click()
-      await $inputsSecondary[1].click()
-
-      // Check the "None" checkbox in the third group
-      await $inputsOther[1].click()
-
-      // Expect the checkboxes in the first and second groups to be unchecked
-      expect(await getProperty($inputsPrimary[2], 'checked')).toBe(false)
-      expect(await getProperty($inputsSecondary[1], 'checked')).toBe(false)
-    })
-
-    it('hides conditional reveals in other groups', async () => {
-      const $conditionalPrimary = await page.$(
-        `[id="${await getAttribute($inputsPrimary[1], 'aria-controls')}"]`
-      )
-
-      // Check the second checkbox in the first group, which reveals additional content
-      await $inputsPrimary[1].click()
-
-      // Assert that conditional content is revealed
-      expect(await isVisible($conditionalPrimary)).toBe(true)
-
-      // Check the "None" checkbox in the third group
-      await $inputsOther[1].click()
-
-      // Assert that the second checkbox in the first group is unchecked
-      expect(await getProperty($inputsPrimary[1], 'checked')).toBe(false)
-
-      // Expect conditional content to have been collapsed
-      expect(await isVisible($conditionalPrimary)).toBe(false)
-    })
-
     describe('errors at instantiation', () => {
       it('throws when GOV.UK Frontend is not supported', async () => {
         await expect(
