+++ packages/govuk-frontend/src/govuk/components/exit-this-page/exit-this-page.puppeteer.test.js
@@ -1,13 +1,12 @@
 const { setTimeout } = require('timers/promises')
 
-const { goToExample, render } = require('@govuk-frontend/helpers/puppeteer')
+const { render } = require('@govuk-frontend/helpers/puppeteer')
 const { getExamples } = require('@govuk-frontend/lib/components')
 
-const buttonClass = '.govuk-js-exit-this-page-button'
-const skiplinkClass = '.govuk-js-exit-this-page-skiplink'
-const overlayClass = '.govuk-exit-this-page-overlay'
-
 describe('/components/exit-this-page', () => {
+  const buttonClass = '.govuk-js-exit-this-page-button'
+  const overlayClass = '.govuk-exit-this-page-overlay'
+
   let examples
 
   beforeAll(async () => {
@@ -43,22 +42,6 @@ describe('/components/exit-this-page', () => {
 
       await page.setRequestInterception(false)
     })
-
-    it('navigates to the href of the skiplink', async () => {
-      const page = await goToExample(browser, 'exit-this-page-with-skiplink')
-
-      const exitPageURL = await page
-        .$eval(skiplinkClass, (el) => el.getAttribute('href'))
-        .then((path) => new URL(path, 'file://'))
-
-      await Promise.all([
-        page.waitForNavigation(),
-        page.focus(skiplinkClass),
-        page.click(skiplinkClass)
-      ])
-
-      expect(page.url()).toBe(exitPageURL.href)
-    })
   })
 
   describe('when JavaScript is available', () => {
@@ -81,56 +64,6 @@ describe('/components/exit-this-page', () => {
       expect(page.url()).toBe(exitPageURL.href)
 
       await page.setRequestInterception(false)
-    })
-
-    it('navigates to the href of the skiplink', async () => {
-      const page = await goToExample(browser, 'exit-this-page-with-skiplink')
-
-      const exitPageURL = await page
-        .$eval(skiplinkClass, (el) => el.getAttribute('href'))
-        .then((path) => new URL(path, 'file://'))
-
-      await Promise.all([
-        page.waitForNavigation(),
-        page.focus(skiplinkClass),
-        page.click(skiplinkClass)
-      ])
-
-      expect(page.url()).toBe(exitPageURL.href)
-    })
-
-    it('shows the ghost page when the EtP button is clicked', async () => {
-      const page = await goToExample(browser, 'exit-this-page-with-skiplink')
-
-      // Stop the button from navigating away from the current page as a workaround
-      // to puppeteer struggling to return to previous pages after navigation reliably
-      await page.$eval(buttonClass, (el) => el.setAttribute('href', '#'))
-      await page.click(buttonClass)
-
-      const ghostOverlay = await page.evaluate(
-        (overlayClass) => document.body.querySelector(overlayClass),
-        overlayClass
-      )
-      expect(ghostOverlay).not.toBeNull()
-    })
-
-    it('shows the ghost page when the skiplink is clicked', async () => {
-      const page = await goToExample(browser, 'exit-this-page-with-skiplink')
-
-      // Stop the button from navigating away from the current page as a workaround
-      // to puppeteer struggling to return to previous pages after navigation reliably
-      //
-      // We apply this to the button and not the skiplink because we pull the href
-      // from the button rather than the skiplink
-      await page.$eval(buttonClass, (el) => el.setAttribute('href', '#'))
-      await page.focus(skiplinkClass)
-      await page.click(skiplinkClass)
-
-      const ghostOverlay = await page.evaluate(
-        (overlayClass) => document.body.querySelector(overlayClass),
-        overlayClass
-      )
-      expect(ghostOverlay).not.toBeNull()
     })
 
     describe('keyboard shortcut activation', () => {
